Fixed a linker error that occured with the now warning-fixed GRand.h.
Ensured that the debug version of wizard/usage.cpp was built for
inclusion in the gcov source.

+++ src/wizard/main.cpp
@@ -966,7 +966,7 @@ void LaunchBrowser(const char* szAddress)
 void do_wizard()
 {
 	int port = 8421;
-	unsigned int seed = getpid() * (unsigned int)time(NULL);
+	uint64_t seed = getpid() * (unsigned int)time(NULL);
 	GRand prng(seed);
 	bool serverStarted = false;
 	try
